Add set_calc_mode() to set calcMode and calcId
This commit introduces the new method set_calc_mode() to allow user to
set calculation mode and calcId explicitly. It enables user to work
around the save file propmt which was reported as issue #100. It also
removes the 'fullCalcOnLoad' attribute and set the 'calcOnSave'
attribute to 0 by default.
The hide colunm using set_column() in 0.77 does not work. The patch is
applied to 0.72 which is the last working version.